Adds the utils folder with common utilities
The folder contains a utility to read plain text files line by line to
return a vector of strings, each representing a line in the file that
has been read.

Common utilities for most days of advent of code can be places in this
folder to be reused by challenges in later days.