Adds utility functions specific to the day 2 challenge
split_line is supposed to act as a C++ equivalent to the python function
split(). Given a string, it will split it into subparts according to a
delimiter character.

process_lines will do the bulk of the work for this challenge. It will
among other things call split_line on every string in the input data.