Adds solution to second challenge of day 2
The solution is very similar to the functionality already present
in the solution for the first challenge of day 2. Since the
unordered_map already contains the information about the minimum
amount of gems of each color needed to play each game, the program
only needed to be modified to compute the power (as specified in
the challenge) for each game and accumulate for all games played.

The solution computed was submitted and the submission passed.
Second challenge of day 2 completed and another gold star awarded!