Day 3: Challenge 1 - Adds helper function to parse numbers
Adds parse_numbers which is supposed to be called every time a
numerical character is encountered while iterating over the input
data.

The function then parses the number it encounters and returns it
together with the index of the the last character that belong to
the number in the input data.