Day 3: Challenge 1 - Solution
Adds helper functions is_part_number and is_valid_index which are
used to determine if a parsed number is a part number.

Fixes a bug present in parse_number which would return the index
of the next non numerical character in the matrix of characters
corresponding to the input.

The solution produced with this solution was submitted and it
passed the submission, meaning the first challenge of day 3 is
complete!