Day 3: Challenge 2 - Adds strcture to .cpp file and changes functionality
The content from the .cpp file od the first challenge of day 3 is copied
over to the .cpp file for the second challenge of day 3. The solution
for the second challenge will be based on the same structure used to
solve the first challenge of the day.

The function is_part_number is modified to check all adjacent
characters to any number that is found. This functionality will allow
assisgning the found number to any gears adjacent to it.