Day 3: Challenge 2 - Solution
Modifies is_part_number to to take the parsed number and a map
containing key-value pairs corresponding to a pair of integers
(the indexes for the current position of the matrix) and a vector
of integers.

This way, since is_part_number now checks all adjacent characters
to a parsed number, whenever it encounters a gear ('*') it simply
need to use its indexes as key for the map and push the parsed
number to the back of the vector contained in the map.

This ensures each number is added to every '*' it is adjacent to.
To solve the problem then you only have to iterate over all vectors
of integers contained in the map when the loop for the first challenge
of day 3 is done. The vectors with length 2 will correspond to actual
gears, and their part number can be computed by multiplying both
numbers in the vector. After that it is simply a matter of aggregating.

The answer that was computed was submitted and it passed the
submission. Day 3 is complete and 6 golden stars have been collected!