Moves split_line into the utils folder for reusability
For the challenges of day 2, the function split_line was implemented.
This function is suposed to be an equivalent to the python split
function to split a string on a given delimiter and return a vector
of substrings produced.

This commit moves this function into the utils folder for resuability
ahead of the challenges of day 4, and probaly for future days as well.