Day 4: Challenge 2 - Adds .cpp file and modifies functionality from first challenge
Adds a .cpp file for the second challenge of day 4. The file is from the beginning
a copy of the solution for the first challenge of day 4. The functionality of it is
sligthly modified to first iterate over all scartch cards and simply make a record
of all winning numbers and numbers of each scratch card. These are stored in a vector
and the index at which they are stored at correspond to the number of the scratch card.

Extracts the functionality to compute the score of a scartch card to a helper function
instead of being in the main function.